requirements/requirements.txt
amqp==5.2.0
asgiref==3.7.2
autopep8==2.0.4
billiard==4.2.0
boto3==1.28.71
botocore==1.31.71
celery==5.3.6
certifi==2023.11.17
channels==4.0.0
channels-redis==4.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
defusedxml==0.7.1
distlib==0.3.7
Django==5.0
django-celery-beat==2.1.0
django-cors-headers==4.3.1
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5
django-nine==0.2.7
django-redis==5.4.0
django-silk==5.0.4
django-timezone-field==4.2.3
django-user-agents==0.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
elastic-transport==8.11.0
elasticsearch==8.11.1
elasticsearch-dsl==8.11.0
et-xmlfile==1.1.0
filelock==3.12.4
gprof2dot==2022.7.29
inflection==0.5.1
jmespath==1.0.1
kombu==5.3.4
msgpack==1.0.7
numpy==1.26.1
openpyxl==3.1.2
packaging==23.2
pandas==2.1.1
Pillow==10.1.0
platformdirs==3.11.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycodestyle==2.11.1
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
redis-simple-cache==0.0.8
s3transfer==0.7.0
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==2.0.7
user-agents==2.2.0
vine==5.1.0
virtualenv==20.24.5
wcwidth==0.2.12
whitenoise==6.6.0


django-plotly-dash==2.3.1
dash-ag-grid==2.4.0
django-crum==0.7.9
dash-bootstrap-components==1.0.0
dash-bootstrap-templates==1.0.8
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-mantine-components==0.12.1
dash-iconify==0.1.2
dash-renderer==1.9.1
dpd-components==0.1.0
dpd-static-support==0.0.5
requests==2.31.0
mysql-connector-python==8.3.0
PyMySQL==1.1.0
flower==2.0.1
cryptography==42.0.5